+++ frontend/src/components/CountryFlag/utils.ts
@@ -10,7 +10,15 @@ export function getAlpha2CodeFromAlpha2or3Code(countryCode: string | undefined):
   }
 
   try {
-    return (countryCode.length === 3 ? countries.alpha3ToAlpha2(countryCode) : countryCode).toLowerCase()
+    if (countryCode.length === 3) {
+      return countries.alpha3ToAlpha2(countryCode).toLowerCase()
+    }
+
+    if (countryCode.length === 2) {
+      return countryCode.toLowerCase()
+    }
+
+    return undefined
   } catch (err) {
     return undefined
   }
