+++ frontend/src/features/PriorNotification/components/PriorNotificationCard/index.tsx
@@ -2,21 +2,24 @@ import { useGetGearsQuery } from '@api/gear'
 import { ErrorWall } from '@components/ErrorWall'
 import { FrontendErrorBoundary } from '@components/FrontendErrorBoundary'
 import { LogbookMessage } from '@features/Logbook/components/VesselLogbook/LogbookMessages/messages/LogbookMessage'
-import { pdfContent, pdfStyle } from '@features/PriorNotification/components/PriorNotificationCard/constants'
+import { HTML_STYLE } from '@features/PriorNotification/components/PriorNotificationCard/template'
+import { pdfContent } from '@features/PriorNotification/components/PriorNotificationCard/utils'
 import { useMainAppDispatch } from '@hooks/useMainAppDispatch'
 import { useMainAppSelector } from '@hooks/useMainAppSelector'
 import { DisplayedErrorKey } from '@libs/DisplayedError/constants'
-import { Accent, Button, Dropdown, Icon } from '@mtes-mct/monitor-ui'
+import { Accent, Button, customDayjs, Dropdown, Icon } from '@mtes-mct/monitor-ui'
 import printJS from 'print-js'
 import { useMemo } from 'react'
 import styled from 'styled-components'
 import { LoadingSpinnerWall } from 'ui/LoadingSpinnerWall'
 
 import { Header } from './Header'
+import { useIsSuperUser } from '../../../../auth/hooks/useIsSuperUser'
 import { priorNotificationActions } from '../../slice'
 
 export function PriorNotificationCard() {
   const dispatch = useMainAppDispatch()
+  const isSuperUser = useIsSuperUser()
   const priorNotificationDetail = useMainAppSelector(state => state.priorNotification.priorNotificationCardDetail)
   const sideWindowPriorNotificationCardError = useMainAppSelector(
     state => state.displayedError.sideWindowPriorNotificationCardError
@@ -41,9 +44,9 @@ export function PriorNotificationCard() {
 
   const downloadPDF = () => {
     printJS({
-      documentTitle: '',
-      printable: pdfContent(priorNotificationDetail.logbookMessage, gearsWithName),
-      style: pdfStyle,
+      documentTitle: `preavis_entree_port_debarquement_${customDayjs().utc().format('DDMMYYYY')}.pdf`,
+      printable: pdfContent(priorNotificationDetail?.logbookMessage, gearsWithName),
+      style: HTML_STYLE,
       type: 'raw-html'
     })
   }
@@ -94,7 +97,7 @@ export function PriorNotificationCard() {
             </Button>
             <Dropdown accent={Accent.PRIMARY} Icon={Icon.Download} placement="topEnd" title="Télécharger les documents">
               <>
-                {priorNotificationDetail.logbookMessage.flagState !== 'FR' && (
+                {isSuperUser && priorNotificationDetail.logbookMessage.flagState !== 'FR' && (
                   <Dropdown.Item onClick={downloadPDF}>
                     Autorisation d&apos;entrée au port et de débarquement
                   </Dropdown.Item>
