+++ frontend/src/features/PriorNotification/components/PriorNotificationCard/types.ts
@@ -1,0 +1,1 @@
+export type TemplateData = { [key: string]: string | number | null | undefined }
